Registry without inheritance

+++ src/sak/aligned_allocator.hpp
@@ -131,14 +131,14 @@ namespace sak
         /// can find the old_ptr again when we need to deallocate the memory
         pointer allocate(size_type num, const void * = 0)
             {
-                uint32_t space_needed = num * sizeof(T) + Alignment;
+                uint32_t space_needed = static_cast<uint32_t>(num * sizeof(T)) + Alignment;
 
                 uint8_t *old_ptr = reinterpret_cast<uint8_t*>(
                     ::operator new(space_needed));
 
                 uint8_t *new_ptr = find_aligned(old_ptr);
 
-                uint32_t offset = new_ptr - old_ptr;
+                uint32_t offset = static_cast<uint32_t>(new_ptr - old_ptr);
 
                 assert(offset > 0);
                 assert(offset <= Alignment);
